(PC-32694)[PRO] refactor: delete spec directories in CollectiveOfferSummary
